pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "patzer"
version = "0.1.0"
description = "Chess improvement tool with game analysis, theme tagging, and spaced repetition"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.49.0",
    "chess>=1.10.0",
    "click>=8.1.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
patzer = "patzer.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/patzer"]

[tool.hatch.build]
sources = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
